requirements.txt
absl-py==0.6.1
astor==0.7.1
bleach==3.0.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dominate==2.3.5
entrypoints==0.2.3
fastcache==1.0.2
flickr-api==0.6.1
future==0.17.1
gast==0.2.0
grpcio==1.14.1
gym==0.10.9
h5py==2.8.0
httplib2==0.12.0
idna==2.7
ipython-genutils==0.2.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mpmath==1.0.0
nbconvert==5.4.0
nbformat==4.4.0
numpy==1.15.4
oauth==1.0.1
oauth2==1.9.0.post1
opencv-python==3.4.4.19
pandocfilters==1.4.2
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
pyglet==1.3.2
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==1.5
pytz==2018.7
PyYAML==3.12
pyzmq==17.1.2
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
simpy==3.0.11
six==1.11.0
sympy==1.3
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
testpath==0.4.2
torch==0.4.1
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
visdom==0.1.8.5
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
wincertstore==0.2
